requirements.txt
fastapi
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.0+cpu
uvicorn
gunicorn
pandas
sqlalchemy
alembic  # Database migrations
psycopg2-binary
openai
sentence-transformers
qdrant-client
langchain
ragas
flask
flask-sock
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv  # Load environment variables from .env file

# Authentication
PyJWT>=2.8.0  # JWT token handling
argon2-cffi>=23.1.0  # Password hashing (OWASP recommended)

# Integrations
twilio  # For Twilio Flex ACD integration and two-way calling
pyannote.audio  # For speaker diarization
deepgram-sdk  # For Deepgram Nova-3 real-time transcription
groq  # Groq API client for Llama 8B inference (Intelligence Gatekeeper)
